gcs_server/requirements-gcs.txt
fastapi
uvicorn
paho-mqtt
wsproto
websockets